+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/pebble/HealthSampleProvider.java
@@ -4,27 +4,27 @@ import nodomain.freeyourgadget.gadgetbridge.devices.SampleProvider;
 import nodomain.freeyourgadget.gadgetbridge.model.ActivityKind;
 
 public class HealthSampleProvider implements SampleProvider {
-    public static final byte TYPE_DEEP_SLEEP = 5;
-    public static final byte TYPE_LIGHT_SLEEP = 4;
-    public static final byte TYPE_ACTIVITY = -1;
+    public static final int TYPE_DEEP_SLEEP = 5;
+    public static final int TYPE_LIGHT_SLEEP = 4;
+    public static final int TYPE_ACTIVITY = -1;
 
     protected final float movementDivisor = 8000f;
 
     @Override
-    public int normalizeType(byte rawType) {
+    public int normalizeType(int rawType) {
         switch (rawType) {
-        case TYPE_DEEP_SLEEP:
-            return ActivityKind.TYPE_DEEP_SLEEP;
-        case TYPE_LIGHT_SLEEP:
-            return ActivityKind.TYPE_LIGHT_SLEEP;
-        case TYPE_ACTIVITY:
-        default:
-            return ActivityKind.TYPE_UNKNOWN;
-    }
+            case TYPE_DEEP_SLEEP:
+                return ActivityKind.TYPE_DEEP_SLEEP;
+            case TYPE_LIGHT_SLEEP:
+                return ActivityKind.TYPE_LIGHT_SLEEP;
+            case TYPE_ACTIVITY:
+            default:
+                return ActivityKind.TYPE_UNKNOWN;
+        }
     }
 
     @Override
-    public byte toRawActivityKind(int activityKind) {
+    public int toRawActivityKind(int activityKind) {
         switch (activityKind) {
             case ActivityKind.TYPE_ACTIVITY:
                 return TYPE_ACTIVITY;
@@ -40,13 +40,13 @@ public class HealthSampleProvider implements SampleProvider {
 
 
     @Override
-    public float normalizeIntensity(short rawIntensity) {
+    public float normalizeIntensity(int rawIntensity) {
         return rawIntensity / movementDivisor;
     }
 
 
     @Override
-    public byte getID() {
+    public int getID() {
         return SampleProvider.PROVIDER_PEBBLE_HEALTH;
     }
 }
